refactor: /conponentsの整理
