+++ pages/components.tsx
@@ -24,8 +24,6 @@ import { YouAre } from "@/components/common/YouAre";
 import { DeleteHintList } from "@/components/DeleteHintList";
 import { Answer } from "@/components/game/Answer";
 import { AnswerWait } from "@/components/game/AnswerWait";
-import { BanishPerson } from "@/components/game/BanishPerson";
-import { ChoiceWolf } from "@/components/game/ChoiceWolf";
 import { DeleteHintOtherMasterUI } from "@/components/game/DeleteHintOtherMasterUI";
 import { DiscussJudgeAns } from "@/components/game/DiscussJudgeAns";
 import { HowToDecideTheme } from "@/components/game/HowToDecideTheme";
@@ -36,6 +34,8 @@ import { Result } from "@/components/game/Result";
 import { SelectDuplicateHint } from "@/components/game/SelectDuplicateHint";
 import { ThinkingTheme } from "@/components/game/ThinkingTheme";
 import { Wait } from "@/components/game/Wait";
+import { BanishPerson } from "@/components/game/wolf/BanishPerson";
+import { ChoiceWolf } from "@/components/game/wolf/ChoiceWolf";
 import { MemberList } from "@/components/MemberList";
 
 import { avatarList } from "@/data/AvatarList";
@@ -44,7 +44,6 @@ const Component: NextPage = () => {
   const [text, setText] = useState("");
   const [nickname, setNickname] = useState("");
   const [avatarIndex, setAvatarIndex] = useState(0);
-  // eslint-disable-next-line no-unused-vars, @typescript-eslint/no-unused-vars
   const [step, setStep] = useState<number>(0);
   const theme = "theme";
   const answer = "answer";
@@ -206,6 +205,49 @@ const Component: NextPage = () => {
             </CardBody>
           </Card>
 
+          {/* common/YouAre */}
+          <VSpacer size={8} />
+          <Heading size="lg">common/YouAre</Heading>
+          <Card variant="filled">
+            <CardBody>
+              <YouAre
+                title="あなたは人狼です！"
+                text="正解しないように妨害をしよう！"
+                areYou={true}
+              />
+            </CardBody>
+          </Card>
+          <VSpacer size={8} />
+          <Heading size="lg">common/YouAre</Heading>
+          <Card variant="filled">
+            <CardBody>
+              <YouAre
+                title="あなたは村人です！"
+                text="協力して正解に導こう！"
+                areYou={false}
+              />
+            </CardBody>
+          </Card>
+
+          {/* DeleteHintList */}
+          <VSpacer size={8} />
+          <Heading size="lg">DeleteHintList</Heading>
+          <Card variant="filled">
+            <CardBody>
+              <DeleteHintList hintList={exampleHintList} />
+              <VSpacer size={8} />
+              <Button
+                colorScheme="blue"
+                onClick={() => {
+                  // eslint-disable-next-line no-console
+                  console.log(exampleHintList);
+                }}
+              >
+                チェックされているものをコンソールで確認
+              </Button>
+            </CardBody>
+          </Card>
+
           {/* game/Answer */}
           <VSpacer size={8} />
           <Heading size="lg">game/Answer</Heading>
@@ -226,15 +268,6 @@ const Component: NextPage = () => {
             </CardBody>
           </Card>
 
-          {/* game/DiscussJudgeAns */}
-          <VSpacer size={8} />
-          <Heading size="lg">game/DiscussJudgeAns</Heading>
-          <Card variant="filled">
-            <CardBody>
-              <DiscussJudgeAns theme={text} answer={text} setStep={setStep} />
-            </CardBody>
-          </Card>
-
           {/* game/DeleteHintOtherMasterUI */}
           <VSpacer size={8} />
           <Heading size="lg">game/DeleteHintOtherMasterUI</Heading>
@@ -245,6 +278,15 @@ const Component: NextPage = () => {
                 setStep={setStep}
               />
               <VSpacer size={8} />
+            </CardBody>
+          </Card>
+
+          {/* game/DiscussJudgeAns */}
+          <VSpacer size={8} />
+          <Heading size="lg">game/DiscussJudgeAns</Heading>
+          <Card variant="filled">
+            <CardBody>
+              <DiscussJudgeAns theme={text} answer={text} setStep={setStep} />
             </CardBody>
           </Card>
 
@@ -263,6 +305,15 @@ const Component: NextPage = () => {
           <Card variant="filled">
             <CardBody>
               <InputHint theme="小三元" setStep={setStep} />
+            </CardBody>
+          </Card>
+
+          {/* game/InputTheme */}
+          <VSpacer size={8} />
+          <Heading size="lg">game/InputTheme</Heading>
+          <Card variant="filled">
+            <CardBody>
+              <InputTheme setStep={setStep} />
             </CardBody>
           </Card>
 
@@ -326,43 +377,9 @@ const Component: NextPage = () => {
           </Card>
           <VSpacer size={12} />
 
-          {/* common/YouAre */}
+          {/* game/wolf/BanishPerson*/}
           <VSpacer size={8} />
-          <Heading size="lg">common/YouAre</Heading>
-          <Card variant="filled">
-            <CardBody>
-              <YouAre
-                title="あなたは人狼です！"
-                text="正解しないように妨害をしよう！"
-                areYou={true}
-              />
-            </CardBody>
-          </Card>
-          <VSpacer size={8} />
-          <Heading size="lg">common/YouAre</Heading>
-          <Card variant="filled">
-            <CardBody>
-              <YouAre
-                title="あなたは村人です！"
-                text="協力して正解に導こう！"
-                areYou={false}
-              />
-            </CardBody>
-          </Card>
-
-          {/* game/ChoiceWolf*/}
-          <VSpacer size={8} />
-          <Heading size="lg">game/ChoiceWolf</Heading>
-          <Card variant="filled">
-            <CardBody>
-              <ChoiceWolf wolfList={exampleWolfHintList} />
-              <p>{text}</p>
-            </CardBody>
-          </Card>
-
-          {/* game/BanishPerson*/}
-          <VSpacer size={8} />
-          <Heading size="lg">game/BanishPerson</Heading>
+          <Heading size="lg">game/wolf/BanishPerson</Heading>
           <Card variant="filled">
             <CardBody>
               <BanishPerson player="KJ" vote={6} />
@@ -370,22 +387,13 @@ const Component: NextPage = () => {
             </CardBody>
           </Card>
 
-          {/* DeleteHintList */}
+          {/* game/wolf/ChoiceWolf*/}
           <VSpacer size={8} />
-          <Heading size="lg">DeleteHintList</Heading>
+          <Heading size="lg">game/wolf/ChoiceWolf</Heading>
           <Card variant="filled">
             <CardBody>
-              <DeleteHintList hintList={exampleHintList} />
-              <VSpacer size={8} />
-              <Button
-                colorScheme="blue"
-                onClick={() => {
-                  // eslint-disable-next-line no-console
-                  console.log(exampleHintList);
-                }}
-              >
-                チェックされているものをコンソールで確認
-              </Button>
+              <ChoiceWolf wolfList={exampleWolfHintList} />
+              <p>{text}</p>
             </CardBody>
           </Card>
 
@@ -398,15 +406,6 @@ const Component: NextPage = () => {
                 title={"参加者リスト"}
                 memberNameList={exampleNameList}
               />
-            </CardBody>
-          </Card>
-
-          {/* game/InputTheme */}
-          <VSpacer size={8} />
-          <Heading size="lg">game/InputTheme</Heading>
-          <Card variant="filled">
-            <CardBody>
-              <InputTheme setStep={setStep} />
             </CardBody>
           </Card>
