+++ src/components/game/wolf/BanishPerson.tsx
